Feat: close #9 close #10 디자인 구현, 폴더 구조 정리
- 현재까지 기능 구현 마친 로그인, 회원가입 페이지 반응형으로 디자인 구현
- 폴더 구조 정리 (services 폴더 생성 <- firebase.tsx 파일, routes 폴더 내 private/public 폴더 생성)
