+++ src/components/ui/drawerRight.tsx
@@ -28,14 +28,9 @@ const DrawerRightContent = React.forwardRef<
     <DrawerRightOverlay />
     <DrawerPrimitive.Content
       ref={ref}
-      className={cn(
-        "bg-white flex flex-col h-full w-[300px] mt-24 fixed bottom-0 right-0"
-        // "fixed inset-x-0 bottom-0 z-50 mt-24 flex h-auto flex-col rounded-t-[10px] border bg-background",
-        // className
-      )}
+      className={cn("bg-white flex flex-col h-full w-[300px] mt-24 fixed bottom-0 right-0")}
       {...props}
     >
-      {/* <div className="p-4 bg-white flex-1 h-full" /> */}
       {children}
     </DrawerPrimitive.Content>
   </DrawerRightPortal>
