+++ src/components/ui/formInput.tsx
@@ -9,7 +9,7 @@ const Input = React.forwardRef<HTMLInputElement, InputProps>(({ className, type,
     <input
       type={type}
       className={cn(
-        "my-0 flex h-full outline outline-1 outline-neutral-400 bg-background px-2 py-1.5 text-sm file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-neutral-800 disabled:cursor-not-allowed disabled:opacity-50",
+        "my-0 flex h-full w-full rounded-sm outline outline-1 outline-zinc-400 bg-background px-2 py-1.5 text-sm file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-zinc-800 disabled:cursor-not-allowed disabled:opacity-50",
         className
       )}
       ref={ref}
