+++ src/context/authContext.tsx
@@ -2,3 +2,4 @@ import { createContext } from "react";
 import { UserInfoType } from "@/types";
 
 export const AuthContext = createContext<UserInfoType | null>(null);
+//
