+++ src/provider/authProvider.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, useEffect, useState } from "react";
 import { AuthContext } from "../context/authContext";
-import { firebaseAuth, getUser } from "@/firebase";
+import { firebaseAuth, getUser } from "@/services/firebase";
 import { UserInfoType } from "@/types";
 //provider/authProvider.tsx 에서 유저 정보를 내려줘야 한다.
 // onAuthStateChanged를 실행하고 useEffect를 사용해 변경을 감지하자.
