--- src/routes/admin/products.tsx
@@ -1,4 +1,0 @@
-export default function Products() {
-  // 판매상품 리스트 목록
-  return <h3>판매자 페이지</h3>;
-}
