+++ src/routes/index.tsx
@@ -17,14 +17,13 @@ import {
   DrawerRightFooter,
   DrawerRightClose,
 } from "@/components/ui/drawerRight";
-import DocsSidebarNav from "@/components/ui/sidebar";
+import DocsSidebarNav from "@/components/sidebar";
 import { Link, Outlet } from "react-router-dom";
 import logo from "src/assets/logo.png";
 import { Input } from "@/components/ui/formInput";
-import { useContext, useEffect, useState } from "react";
-import { getAuth, onAuthStateChanged } from "firebase/auth";
-import { getUser } from "@/firebase";
+import { useContext } from "react";
 import { AuthContext } from "@/context/authContext";
+import { HeaderNavItem } from "@/types";
 
 export const sidebarNav = [
   {
@@ -61,24 +60,27 @@ export default function Layout() {
   const userInfo = useContext(AuthContext);
   const isAdmin = userInfo?.type === "관리자" ? true : false;
   // 전역관리 유저타입 저장
-  const headerNav = [
+  const headerNav: HeaderNavItem[] = [
     userInfo ? { title: "로그아웃", href: "/" } : { title: "로그인", href: "/login" },
     { title: "장바구니", href: "/basket" },
-    !userInfo && { title: "회원가입", href: "/signup" },
+    !userInfo ? { title: "회원가입", href: "/signup" } : { title: "", href: "" },
     !isAdmin ? { title: "마이페이지", href: "/mypage" } : { title: "주문조회", href: "/orders" },
     // isAdmin && { title: "주문조회", href: "/orders" },
   ];
 
   return (
-    <div className="w-full flex flex-col h-screen bg-slate-300">
+    <div className="w-full flex flex-col h-screen">
+      {/* 데스크탑 레이아웃 */}
       <div className="hidden md:inline-block">
         <Header items={headerNav} />
-        <div className="hidden md:flex flex-row bg-slate-100 flex-grow">
-          <Link to={"/"} className="flex flex-col">
-            <img src={logo} width={120} />
+        <div className="px-12 py-8 hidden md:flex flex-row flex-grow h-full">
+          <div>
+            <Link to={"/"} className="flex flex-col mb-7">
+              <img src={logo} width={120} />
+            </Link>
             <DocsSidebarNav items={sidebarNav} />
-          </Link>
-          <div className="p-10 flex-grow">
+          </div>
+          <div className="flex-grow ">
             <Outlet />
           </div>
         </div>
@@ -87,31 +89,8 @@ export default function Layout() {
       {/* 모바일 레이아웃 */}
       <div className="md:hidden">
         <Header items={headerNav} />
-        <div className="flex justify-between">
-          <DrawerLeft direction="left">
-            <DrawerLeftTrigger>
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="h-8 w-8"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 12h16M4 18h16" />
-              </svg>
-            </DrawerLeftTrigger>
-            <DrawerLeftContent>
-              <DrawerLeftHeader>
-                <DrawerLeftTitle>
-                  <DocsSidebarNav items={sidebarNav} />
-                </DrawerLeftTitle>
-                {/* <DrawerLeftDescription>This action cannot be undone.</DrawerLeftDescription> */}
-              </DrawerLeftHeader>
-              <DrawerLeftFooter>
-                <DrawerLeftClose>{/* <Button variant="outline">Cancel</Button> */}</DrawerLeftClose>
-              </DrawerLeftFooter>
-            </DrawerLeftContent>
-          </DrawerLeft>
+        <div className="px-5 py-3 flex justify-between">
+          <DocsSidebarNav items={sidebarNav} />
           <Link to={"/"}>
             <img src={logo} width={70} />
           </Link>
@@ -145,7 +124,9 @@ export default function Layout() {
             </DrawerRightContent>
           </DrawerRight>
         </div>
-        <Outlet />
+        <div className="flex-grow h-full px-6">
+          <Outlet />
+        </div>
       </div>
     </div>
   );
