+++ src/routes/private/admin/addProduct.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function AddProduct() {
+  return <div>상품 등록 화면</div>;
+}
