+++ src/routes/private/admin/index.tsx
@@ -1,7 +1,8 @@
 // import { lazy } from "react";
 import { Navigate } from "react-router-dom";
-import Orders from "./admin/orders";
-import Products from "./admin/products";
+import Orders from "./orders";
+import Products from "./products";
+import AddProduct from "./addProduct";
 
 // 로그인 한 상태에서 판매자(관리자)가 볼 수 있는 화면
 export default function PrivateRoutes() {
@@ -11,7 +12,7 @@ export default function PrivateRoutes() {
       { path: "/orders", element: <Orders /> }, // 주문조회
       { path: "/login", element: <Navigate to="/" replace /> }, // 로그인 했으면 들어가지 못함.
       { path: "/signup", element: <Navigate to="/" replace /> }, // 로그인 했으면 들어가지 못함. 너무 느림
-
+      { path: "/add-product", element: <AddProduct /> },
       // { path: "*", element: <Navigate to="/" replace /> },
     ],
   };
