+++ src/routes/public/signup.tsx
@@ -1,4 +1,4 @@
-import SignUpForm from "@/components/ui/signUpForm";
+import SignUpForm from "@/components/signUpForm";
 import { useNavigate } from "react-router-dom";
 
 export default function SignUp() {
@@ -6,6 +6,7 @@ export default function SignUp() {
   return (
     <div>
       <h3 className="text-xl">회원가입</h3>
+      <hr className="mt-5" />
       <div className="mt-8">
         <SignUpForm navigate={navigate} />
       </div>
